requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.3            # via black
asgiref==3.2.7            # via django
attrs==19.3.0             # via black, pytest
black==19.10b0            # via -r requirements.in
click==7.1.1              # via black, pip-tools
django-debug-toolbar==2.2  # via -r requirements.in
django==3.0.5             # via -r requirements.in, django-debug-toolbar, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pathspec==0.8.0           # via black
pip-tools==5.0.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.5    # via -r requirements.in
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django
pytz==2019.3              # via django
regex==2020.4.4           # via black
six==1.14.0               # via packaging, pip-tools
sqlparse==0.3.1           # via django, django-debug-toolbar
toml==0.10.0              # via black
typed-ast==1.4.1          # via black
wcwidth==0.1.9            # via pytest

# The following packages are considered to be unsafe in a requirements file:
# pip
